chore(BUX-164): move BEEF tx SPV to separate file
