+++ server/p2p_receive_transaction.go
@@ -87,7 +87,7 @@ func (c *Configuration) p2pReceiveBeefTx(w http.ResponseWriter, req *http.Reques
 		panic("empty beef after parsing!")
 	}
 
-	err := dBeef.ExecuteSimplifiedPaymentVerification(c.actions)
+	err := paymail.ExecuteSimplifiedPaymentVerification(dBeef, c.actions)
 	if err != nil {
 		ErrorResponse(w, ErrorSimplifiedPaymentVerification, err.Error(), http.StatusExpectationFailed)
 		return
